feat : 1106 N:1 - 02 실습완료

+++ code/1weeks/1106/02-02-many-to-one-relationships/articles/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
+from django.views.decorators.http import require_http_methods
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
 
@@ -27,11 +28,14 @@ def detail(request, pk):
 
 
 @login_required
+@require_http_methods(['GET','POST'])
 def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -44,8 +48,8 @@ def create(request):
 @login_required
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-
-    article.delete()
+    if request.user == article.user:
+        article.delete()
 
     return redirect('articles:index')
 
@@ -53,16 +57,17 @@ def delete(request, pk):
 @login_required
 def update(request, pk):
     article = Article.objects.get(pk=pk)
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article.pk)
 
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article.pk)
-
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
-
+        return redirect('articles:index')
     context = {
         'article': article,
         'form': form,
@@ -83,6 +88,7 @@ def comments_create(request, pk):
         comment = comment_form.save(commit=False)
         # 외래 키 데이터를 할당
         comment.article = article
+        comment.user = request.user
         comment.save()
         return redirect('articles:detail', article.pk)
     context = {
@@ -97,6 +103,7 @@ def comments_delete(request, article_pk, comment_pk):
     # 삭제할 댓글 조회
     comment = Comment.objects.get(pk=comment_pk)
     # 댓글 삭제
-    comment.delete()
+    if comment.user== request.user:
+        comment.delete()
     # 삭제 후 게시글 상세 페이지로 리다이렉트
     return redirect('articles:detail', article_pk)
